refactor: seperation of concern on controlling layer and adding, remove unused shit
